# Reimplement the default object pool for improved performance