@BrennanConroy 2022-11-23
Oh wait, this is ObjectPool, that wouldn't affect Kestrel specific benchmarks