@adityamandaleeka 2022-11-29
Benchmark code from the zip:

```C#
// © Microsoft Corporation. All rights reserved.

using System.Collections.Generic;
using System.Threading.Tasks;
using BenchmarkDotNet.Attributes;
using Microsoft.Extensions.ObjectPool;

namespace Microsoft.R9.Extensions.Pools.Bench;

#pragma warning disable R9A038 // Use 'Microsoft.R9.Extensions.Pools.ScaledObjectPool' instead of 'Microsoft.Extensions.ObjectPool.DefaultObjectPool' for improved performance
#pragma warning disable S109 // Magic numbers should not be used
#pragma warning disable R9A041 // Use concrete types when possible for improved performance

[MemoryDiagnoser]
public class Pools
{
    private ObjectPool<Foo> _defaultPool = null!;
    private ObjectPool<Foo> _scaledPool = null!;

    [Params(2, 8, 16, 32, 256, 1024, 2048)]
    public int MaxRetained { get; set; }

    [Params(0, 8, 16)]
    public int OverflowCount { get; set; }

    [GlobalSetup]
    public void GlobalSetup()
    {
        _defaultPool = new DefaultObjectPool<Foo>(new DefaultPooledObjectPolicy<Foo>(), MaxRetained);
        _scaledPool = PoolFactory.CreatePool<Foo>(new DefaultPooledObjectPolicy<Foo>(), MaxRetained);

        for (int i = 0; i < MaxRetained; i++)
        {
            _defaultPool.Return(new Foo());
            _scaledPool.Return(new Foo());
        }

    }

    [Benchmark]
    public void DefaultPool()
    {
        UsePool(_defaultPool);
    }

    [Benchmark]
    public void ScaledPool()
    {
        UsePool(_scaledPool);
    }

    [Benchmark]
    public void DefaultPoolThreaded()
    {
        var t0 = Task.Run(() => UsePool(_defaultPool));
        var t1 = Task.Run(() => UsePool(_defaultPool));

#pragma warning disable VSTHRD002 // Avoid problematic synchronous waits
        Task.WaitAll(t0, t1);
#pragma warning restore VSTHRD002 // Avoid problematic synchronous waits
    }

    [Benchmark]
    public void ScaledPoolThreaded()
    {
        var t0 = Task.Run(() => UsePool(_scaledPool));
        var t1 = Task.Run(() => UsePool(_scaledPool));

#pragma warning disable VSTHRD002 // Avoid problematic synchronous waits
        Task.WaitAll(t0, t1);
#pragma warning restore VSTHRD002 // Avoid problematic synchronous waits
    }

    private void UsePool(ObjectPool<Foo> pool)
    {
        var list = new List<Foo>(MaxRetained + OverflowCount);

        for (int i = 0; i < MaxRetained + OverflowCount; i++)
        {
            var foo = pool.Get();
            list.Add(foo);
        }

        foreach (var f in list)
        {
            pool.Return(f);
        }
    }

    private sealed class Foo
    {
    }
}

```