@halter73 2022-11-29
It would be good to check the microbenchmark into a `perf/Microbenchmarks` directory like [the middleware one here](https://github.com/dotnet/aspnetcore/tree/2c0b763e9b5c5f9d79fa1ff5fb9a193c76157329/src/Middleware/perf/Microbenchmarks) as part of this PR. That will make it easier to comment on too.

I think it would probably make sense to test with more threads concurrently getting and then immediately returning pooled objects. As it stands, I don't think this test is creating as much contention as a high-scale web application with objects being both added and removed simultaneously. I think in this more realistic scenario, using the most-recently returned item (i.e. `_firstItem` in the old code) is a lot more important for maintaining cache locality. I also wonder if the list allocations are having an undue impact on the result.

We should figure out what scale we want to focus on. `DefaultObjectPoolProvider.MaximumRetained` defaults to `Environment.ProcessorCount * 2` currently. I think a lot of VMs are running on 8 or less vCPUs which would mean retaining 16 or less objects in the pool. Maybe @sebastienros has a better idea what VM sizes are most popular for ASP.NET Core apps on Azure though. Of course, even if we focus on less retained objects, I don't think that means we should have orders of magnitude worse performance with more retained objects. We might have to look at switching strategies based on the retained size.

We've discussed moving a similar type to the runtime (see https://github.com/dotnet/runtime/issues/49680), but I don't think that should stop us from improving the existing type.