@geeknoid 2022-11-29
@halter73 We run our pools usually with 1024 or 2048 entries so that we can recover from spiky traffic efficiently (if you imagine traffic spiking at 1K RPS and sometimes dropping to <100 RPS). Thus in our environment, it's not bound to the number of processors in the machine, it's bound to the number of concurrent operations that may need one of the pooled resources. That is, we use these pools not so much to avoid repeated allocations of the same object while processing a single request, but more to avoid allocations across multiple incoming concurrent requests.

Using a stack would likely be a net improvement over using a bag for this case, but alas ConcurrentStack allocates and is thus ineffective for this use. But the stack only has a significant effect for quick get/put cases in quick succession, and shouldn't matter nearly as much when dealing with the broader cross-request cases as described above.

 Let me run the test with more contention, although I suspect the results will be very similar. The existing pool implementation is not designed for anything but very small retained sets.